requirements.txt
Flask==2.3.3
nltk==3.8.1
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
chardet==5.2.0